cdctest: move common cdc test infrastructure to a new package
This commit is pure code movement, all fixups happen in the next commit.

Release note: None
